Add files via upload
This is the dictionary of Pali words from the DPD embedded in the mobi version dated November 27th 2023. Contains 75,965 entries after cleaning anything other than Pali and definitions. The format has been left unchanged at this stage but I'll be tweaking the format to comply with the existing dictionary in TBW (escaping characters, handling multiple definitions of the one word etc.)